+++ src/models.rs
@@ -44,8 +44,91 @@ pub trait Model {
         di: &mut DI,
         delay: &mut DELAY,
         options: &ModelOptions,
-    ) -> Result<SetAddressMode, DI::Error>
+    ) -> Result<SetAddressMode, ModelInitError<DI::Error>>
     where
         DELAY: DelayNs,
         DI: Interface;
+}
+
+/// Error returned by [`Model::init`].
+///
+/// This error type is used internally by implementations of the [`Model`]
+/// trait.
+pub enum ModelInitError<DiError> {
+    /// Error caused by the display interface.
+    Interface(DiError),
+
+    /// Invalid configuration error.
+    ///
+    /// This error is returned when the configuration passed to the builder is
+    /// invalid. For example, when the combination of bit depth and interface
+    /// kind isn't supported by the selected model.
+    InvalidConfiguration,
+}
+
+impl<DiError> From<DiError> for ModelInitError<DiError> {
+    fn from(value: DiError) -> Self {
+        Self::Interface(value)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use embedded_graphics::pixelcolor::Rgb565;
+
+    use crate::{
+        Builder,
+        _mock::{MockDelay, MockDisplayInterface},
+        interface::InterfaceKind,
+        InitError,
+    };
+
+    use super::*;
+
+    struct OnlyOneKindModel(InterfaceKind);
+
+    impl Model for OnlyOneKindModel {
+        type ColorFormat = Rgb565;
+
+        const FRAMEBUFFER_SIZE: (u16, u16) = (0, 0);
+
+        fn init<DELAY, DI>(
+            &mut self,
+            _di: &mut DI,
+            _delay: &mut DELAY,
+            _options: &ModelOptions,
+        ) -> Result<SetAddressMode, ModelInitError<DI::Error>>
+        where
+            DELAY: DelayNs,
+            DI: Interface,
+        {
+            if DI::KIND != self.0 {
+                return Err(ModelInitError::InvalidConfiguration);
+            }
+
+            Ok(SetAddressMode::default())
+        }
+    }
+
+    #[test]
+    fn test_assert_interface_kind_serial() {
+        Builder::new(
+            OnlyOneKindModel(InterfaceKind::Serial4Line),
+            MockDisplayInterface,
+        )
+        .init(&mut MockDelay)
+        .unwrap();
+    }
+
+    #[test]
+    fn test_assert_interface_kind_parallel() {
+        assert!(matches!(
+            Builder::new(
+                OnlyOneKindModel(InterfaceKind::Parallel8Bit),
+                MockDisplayInterface,
+            )
+            .init(&mut MockDelay),
+            Err(InitError::InvalidConfiguration)
+        ));
+    }
 }