+++ src/models/ili9486.rs
@@ -3,11 +3,10 @@ use embedded_hal::delay::DelayNs;
 
 use crate::{
     dcs::{BitsPerPixel, PixelFormat, SetAddressMode},
-    interface::Interface,
+    interface::{Interface, InterfaceKind},
+    models::{ili948x, Model, ModelInitError},
     options::ModelOptions,
 };
-
-use super::{ili948x, Model};
 
 /// ILI9486 display in Rgb565 color mode.
 pub struct ILI9486Rgb565;
@@ -24,11 +23,18 @@ impl Model for ILI9486Rgb565 {
         di: &mut DI,
         delay: &mut DELAY,
         options: &ModelOptions,
-    ) -> Result<SetAddressMode, DI::Error>
+    ) -> Result<SetAddressMode, ModelInitError<DI::Error>>
     where
         DELAY: DelayNs,
         DI: Interface,
     {
+        if !matches!(
+            DI::KIND,
+            InterfaceKind::Parallel8Bit | InterfaceKind::Parallel16Bit
+        ) {
+            return Err(ModelInitError::InvalidConfiguration);
+        }
+
         delay.delay_us(120_000);
 
         let pf = PixelFormat::with_all(BitsPerPixel::from_rgb_color::<Self::ColorFormat>());
@@ -45,11 +51,18 @@ impl Model for ILI9486Rgb666 {
         di: &mut DI,
         delay: &mut DELAY,
         options: &ModelOptions,
-    ) -> Result<SetAddressMode, DI::Error>
+    ) -> Result<SetAddressMode, ModelInitError<DI::Error>>
     where
         DELAY: DelayNs,
         DI: Interface,
     {
+        if !matches!(
+            DI::KIND,
+            InterfaceKind::Serial4Line | InterfaceKind::Parallel8Bit | InterfaceKind::Parallel16Bit
+        ) {
+            return Err(ModelInitError::InvalidConfiguration);
+        }
+
         delay.delay_us(120_000);
 
         let pf = PixelFormat::with_all(BitsPerPixel::from_rgb_color::<Self::ColorFormat>());